bindings: python: silence UP045 linter warnings
Linter check UP045 [0] was recently stabilized in ruff 0.12 [1].

Currently, this impacts files that perform:

```
from __future__ import annotations
```

These files are being flagged to use the `X | Y` syntax since this clause
drives the Python interpreter to store type annotations as strings and
defers their evaluation [2] which makes the new syntax "safe".

See also [3] for a brief history.

Similar to UP007 [4], ignore UP045 until the bindings minimally target
Python 3.10.

This is done for the sake of consistency so there is no mixed usage of
`Union[X, Y]` and `Optional[X]` vs the `X | Y` syntax in the bindings.

[0]: https://docs.astral.sh/ruff/rules/non-pep604-annotation-optional/
[1]: https://github.com/astral-sh/ruff/commit/de4fc5b171d8b4a9dfd44ef9b551f1257661ab60
[2]: https://peps.python.org/pep-0563/#implementation
[3]: https://peps.python.org/pep-0649/#a-history-of-annotations
[4]: https://docs.astral.sh/ruff/rules/non-pep604-annotation-union/

Signed-off-by: Vincent Fazio <vfazio@xes-inc.com>