bindings: python: chip: track line IDs during config iteration
Previously, when Chip.request_lines constructed a LineRequest, a mapping
of offset -> line names was maintained for lines requested by name so a
list of identifiers from the config could be generated. This was done by
iterating through offsets and checking to see if a name had been mapped
for the offset. Both the list of line identifiers and the offset map
were then assigned to private attributes of the LineRequest.

Since Chip.request_lines checks for duplicates within the config, the
identifiers within a config are guaranteed to be mutually exclusive. As
such, a line requested by offset cannot also be requested by name and
vice versa.

Now, the list of identifiers is maintained as lines from the config are
iterated which obviates the need to iterate through the list of offsets.
The offset map is no longer maintained or assigned to the LineRequest as
it is not used internally for any purpose.