bindings: python: map names for lines requested by offset
Previously, if a line was requested by name, it could be reconfigured
via its offset, but lines requested by offset could not be reconfigured
via the line name.

Now, the name map built and given to the LineRequest includes the names
for all lines regardless of if it was requested by name or offset.

This allows lines requested by offset to be subsequently reconfigured
via the line name.