bindings: python: warn on unknown lines when reconfiguring
Previously, trying to reconfigure a line that was not in the original
request would be silently ignored, giving no indication to the caller
that an invalid value was specified.

Other public interfaces, such as {get,set}_value[s], raise errors when
invalid values are passed in. However, raising an exception would be a
breaking change for existing users of the reconfigure_lines function.

Instead, issue warnings that an invalid value has been specified to
alert users of the problem.

Signed-off-by: Vincent Fazio <vfazio@gmail.com>